requirements.txt
numpy==1.12.0
Pillow==4.0.0
SoundFile==0.8.1